types: fix webpack-dev-middleware warnings
