+++ packages/shared/package.json
@@ -59,10 +59,6 @@
       "types": "./compiled/loader-utils2/index.d.ts",
       "default": "./compiled/loader-utils2/index.js"
     },
-    "./webpack-dev-middleware": {
-      "types": "./compiled/webpack-dev-middleware/types/index.d.ts",
-      "default": "./compiled/webpack-dev-middleware/index.js"
-    },
     "./webpack-bundle-analyzer": {
       "types": "./compiled/webpack-bundle-analyzer/index.d.ts",
       "default": "./compiled/webpack-bundle-analyzer/index.js"
@@ -163,7 +159,7 @@
     "postcss-value-parser": "4.2.0",
     "prebundle": "1.1.0-beta.6",
     "resolve-url-loader": "5.0.0",
-    "rslog": "^1.2.1",
+    "rslog": "^1.2.2",
     "sass": "^1.75.0",
     "sass-loader": "14.2.1",
     "semver": "^7.6.0",
@@ -173,7 +169,6 @@
     "webpack": "^5.91.0",
     "webpack-bundle-analyzer": "^4.10.2",
     "webpack-chain": "npm:webpack-5-chain@8.0.1",
-    "webpack-dev-middleware": "7.2.1",
     "webpack-merge": "5.10.0",
     "yaml": "^2.4.1"
   },
